Player & Scene Update
I have decided that I have created enough features to warrent keeping a repo, to protect it from my future meddling.

So far I have gotten a 2d character controller from the Unity store (which I have implemented just as it came).

I have also obtained a screen frosting effect from the unity store, I intend to attach this to the weather system in my game.

I have created a method of moving between scenes in which the player not only switches scenes but is also moved into the correct location before the user is aware of the change.