Added - Stats System
I have created a system which allows interaction based on various stats.

The Player has a health, and body temperature stat. If the body temperature falls to low the player's health begins to decrease.

The world has a temperature stat. This stat effects the players body temperature.

The suit (equip to player) has a power state and a climate control stat. the climate control stat bolsters the player's body temperature stat against the world temperature stat. When the player is using their climate control features it drains the suit's battery stat until finally if the battery is empty the climate control turns off.