Quick Fix - Persistant Stats/UI Clipping
I have compiled the three player/suit stats which persist from scene to scene into one scriptable object so that I could remove an unnecessary script from the project. This will hopefully make the flow of the scripts easier to follow.

An issue where the UI was clipping behind the 2D level environment has been fixed. Assigning a camera to display only the UI and making its depth greater than that of the scene’s main camera, has allowed the UI to flow seamlessly over all the objects in the scene.