Added - Universal Render Pipeline
As light has always been a mechanic I had intended to use I found a way to create a dark scene which can be light by a togglable light.

If the light is on it adds a +1 draw on the power supply increase the rate at which it drains.

If the light is off there is no drain to the power supply